Rooms BackEnd #1

+++ WebsiteBookingHotel/admin/ajax/features_facilities.php
@@ -40,9 +40,15 @@ if (isset($_POST['rem_feature'])) {
     $frm_data = filteration($_POST);
     $values = [$frm_data['rem_feature']];
 
-    $q = "DELETE FROM `features` WHERE `id`=?";
-    $res = delete($q, $values, 'i');
-    echo $res;
+    $check_q = select('SELECT * FROm `room_features` WHERE `features_id`=? ',[$frm_data['rem_features']],'i');
+    if(mysqli_num_rows($check_q)==0){
+        $q = "DELETE FROM `features` WHERE `id`=?";
+        $res = delete($q, $values, 'i');
+        echo $res;
+    }else{
+        echo 'room_added';
+    }
+
 }
 
 if (isset($POST['add_facility'])) {
@@ -90,18 +96,22 @@ if (isset($_POST['get_facilites'])) {
 if (isset($_POST['rem_facility'])) {
     $frm_data = filteration($_POST);
     $values = [$frm_data['rem_facility']];
-
-    $pre_q = "SELECT * FROM `facilities` WHERE `id` =?";
-    $res = select($pre_q, $values, 'i');
-    $img = mysqli_fetch_assoc($res);
-
-    if (deleteImage($img['icon'], FACILITIES_FOLDER)) {
-        $q = "DELETE FROM `facilities` WHERE `id`=?";
-        $res = delete($q, $values, 'i');
-        echo $res;
+    $check_q = select('SELECT * FROm `room_facilities` WHERE `facilities_id`=? ',[$frm_data['rem_facilities']],'i');
+    if(mysqli_num_rows($check_q)==0){
+        $pre_q = "SELECT * FROM `facilities` WHERE `id` =?";
+        $res = select($pre_q, $values, 'i');
+        $img = mysqli_fetch_assoc($res);
+    
+        if (deleteImage($img['icon'], FACILITIES_FOLDER)) {
+            $q = "DELETE FROM `facilities` WHERE `id`=?";
+            $res = delete($q, $values, 'i');
+            echo $res;
+        }
+        else{
+            echo 0;
     }
-    else{
-        echo 0;
+    }else{
+        echo 'room_added';
     }
 }
 ?>
